Renamed done conditioner to done component
